Kafka System/producer_line_3/requirements.txt
kafka-python==2.0.2
pip==21.2.4
pymongo==3.12.0
pyparsing==2.4.7
requests
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0